15722 Announce weekday for selectable dates
Previously, the weekday headers were individually selectable accessibility elements, which were not linked in the VoiceOver heirarchy with the DateCells which they related to.
This meant that in VoiceOver, there was no way to know what day of the week the focused day was, and the user had to swipe through the unusable day of the week headings before getting to the DateCells.

This change removes the weekday headings from the VoiceOver heirarchy, and announces the weekday along with the date when the user focuses on a DateCell, e.g. "Wednesday, 12th January 2022".