15722 Calendar a11y: Announce selected dates
Previously, there were no indicators to VoiceOver users that dates were selectable, nor which dates were selected.

Announces date cells as buttons, with those at the beginning and end of of a range, or individually selected, annouced as "Selected".

e.g. VoiceOver would announce "Selected, Thursday, the sixth of January, 2022, button" for a selected date, or "Friday, the seventh of January, 2022, button" for an unselected date.