17780 Display Calendar in self-sized VC
Previously, we displayed the Calendar or UIDatePicker in a half-height sheet on phones. This was a very cramped UI on smaller devices, and also broke the layout of UIDatePicker.

This change repurposes HalfScreenPresentationController, allowing the height of the overlay to be determined by the contained view, fixing the aforementioned layout issues.