17780 Schedule with UIDatePicker on Post Settings
Post Settings also uses a calendar to change the scheduled date for a post. This is presented from a table view, rather than a BottomDrawer, though it is still presented in a partial screen container view.

This change makes Post Settings use a UIDatePicker (on iOS 14+) too, sharing the presentation logic with the PrepublishingViewController.

To achieve that, I've repurposed the SchedulingViewControllerPresenter as PresentableSchedulingViewControllerProvider: it's now just responsible for providing the view controller in a ready-to-present form. Additionally, it supports more customization of the popover source and transitioningDelegate, as required by PublishSettingsViewController.