Bump version number to `19.0.0.5`
I did this to track the fact that we will attempt a new submission after
the `19.0.0.4` one failed because of the App Store icon including the
alpha channel, which should be fixed by
8d09f32a22f21d58b0f9cb7f135d7f25f782799d.

Also notice that I bumped the internal version to "...22", which
technically would mean the build was made on the 22nd of January, even
though (as far as UTC goes) we're still on the 21st. Pragmatically, that
has little consequence because we'll most likely won't be making any
build tomorrow.