Refactor: build My Site screen programatically.
This fixes an issue where the large nav title was not collapsing even if the MSD feature flag was disabled.

The previous My Site screen layout was built via the InterfaceBuilder, and the root view was a stack view.

For the large title to collapse, the root view has to be a scrolling view (table view, collection view, scroll view).

By building the layout programmatically instead of via the IB, we can build two different layouts depending on the flag:

- MSD flag false: root view is the child vc, and the child vc’s root view is a table view
- MSD flag true: root view is stack view for now, which will be updated to a scroll view later