Fix: segmented control logic
The segmented control should be hidden if the blog is NOT a WP.com/Atomic/Jetpack site, or if the device is an iPad.

Also, when switching blogs from a WP.com/Atomic/Jetpack site to a self-hosted site (without Jetpack) and if the dashboard is showing, we need to toggle the segmented control to display the site menu.