Fix the no results view flickering after network call
When updating the no results view, it was removing itself from the view
and then adding it back. This was causing a flicker due to being called
multiple times when switching filters and when the network calls would
finish.

This changes it to only remove the view when there are results returned.