Remove unused `fastlane-plugin-test_center`
The plugin became unused in 75f87334ee, but at the time I forgot to
remove its import from the `Pluginfile`.

I noticed this because, after upgrading to Fastlane 2.203.0, we saw an
error which is apparently due to a name-clash between the `trainer` gem
the plugin requires and the `trainer` Fastlane component.

See:

- Failing build on CircleCI https://app.circleci.com/pipelines/github/wordpress-mobile/WordPress-iOS/27160/workflows/4b4839c8-9d33-4507-afa2-57644d03bc67/jobs/64957
- Failing build on Buildkite https://buildkite.com/wordpress-mobile/wordpress-ios/builds/4716#f3a4719e-6496-4a1e-8cca-051d3b9f083d/664-2569
- Fastlane issue https://github.com/fastlane/fastlane/issues/19834