Replace 🧪 with 🔬 as the Buildkite test step emoji
It's been pointed out that the green color could be confused for a sign
of success for the step, in particular when the observer is not
appropriately caffeinated.

Unfortunately, green is the only available color for the test tube
emoji. I decide to use `:microscope:` to represent the action of looking
at the individual facets of the app's behavior.

Fun fact: Buildkite has a page where it shows [all the emojis it
supports](https://github.com/buildkite/emojis#emoji-reference), standard
and custom.