Remove copy from editorialized release notes about comments moderation
This is a cherry-pick of #17860 (80bc0a1a9a4a3e6bef8b0b6f0c8c15aefd1684f5).

The PR was meant for the `release/19.1` branch, but then update to
target `trunk` to make the process of getting the copy up to GlotPress
for translation smoother.

We need this here, on the `release/19.1` branch, too because those edits
are used to generate the release notes that got into the GitHub release.

Besides, there would be conflicts when merging back into `trunk`
otherwise, too.