Refactor: remove table header view from BlogDetailsVC
We can now remove the table header view from BlogDetailsViewController, because we have extracted the code to SitePickerViewController and we are showing SitePickerVC regardless of the MSD feature flag state.

As such, we can also removed the  BlogDetailHeaderViewDelegate implementation as well as the Remove Site Icon Update Management related code, since these have been implemented on the SitePickerViewController.