Stop logging MediaHost errors to Sentry.
These errors are just an indication that we can't load a site icon
for a private site. It's not a big issue, and we'll proceed to try and
load the icon as if it was a public site. Worst case, we won't display
an icon for the site in question.