Make `release-build-*` Buildkite scripts executable
After adding a parameter to the `release-build-wordpress.sh` call, I got
a permission denied error. See
https://buildkite.com/automattic/wordpress-ios/builds/4907#226a7e18-7592-4e95-9064-2829b4cb7f12/381-383

I guess if the path to a script is the only value in the `command` node,
then Buildkite calls it via `sh` (or maybe `$SHELL`), otherwise it runs
it as an actual command within its shell, in which case if the script is
not executable, it fails.