Check if upload failed when its thumbnail is ready
When an upload's thumbnail is ready, we don't currently check whether the upload actually failed. That leads to a situation where failed uploads are assigned an "uploading" state, which prevents Gutenberg from getting the necessary signals to display an error message for the failure.

With this commit, a check for whether an upload failed is added to prevent this issue.