Add localizable name for WordPress app
Notice that the update to `AppStoreStrings.po` was manual because our
automation is currently unable to add new keys to the `.po` file. See
also:
https://github.com/wordpress-mobile/WordPress-iOS/pull/18116#discussion_r823573389

Also notice that I updated the key used in the `Jetpack-Fastfile` to be
the same as the WordPress one, that is: `app_store_name`. It's safe to
do this update because the value was _never used_ since it wasn't
originally in the Jetpack `.po` file.

I decided not to update the Jetpack logic because I noticed the name
used in `fastlane/appstoreres/jetpack_metadata/source/app_name.txt`,
"Jetpack: WP Speed and Security" is different from what's currently in
the App Store, "Jetpack". I want to consult with the Jetpack and Growth
folks before making any change.