Track `FBReactNativeSpec` hash change after fresh `pod install` run
With "fresh" I mean the following:

```
pod cache clean --all \
  && rm -rf Pods/ \
  && bundle exec pod repo remove trunk \
  && bundle exec pod install
```

This rectifies the change introduced by
437573bdfbaef73d648d8f3604738401504c63dc, of which I was the author.

I didn't notice the change in the diff at the time. The issue was
brought to my attention by @ScoutHarris and @guarani. Thanks folks!
Given you both got the change in this PR and that I was able to
reproduce the `Podfile.lock` not changing locally, then changing after a
fresh install, I think we can file this as an issue with the cache on my
machine.