Fix: show primary vc when collapsing split vc
This fixes an issue where the secondary vc was being shown if the app was launched in landscape orientation then rotated to portrait.

If the user taps an item in the primary vc pane, then we should show the secondary vc when collapsing the split vc. Otherwise, we want to show the primary vc.

# Conflicts:
#	WordPress/Classes/ViewRelated/Blog/Blog Details/BlogDetailsViewController.m