Remove unused variable from `Fastfile`
A leftover of the previous way we did `.strings` extraction from the
file downloaded from GlotPress into the `InfoPlist.strings` and app
extension `.strings` files.

Notice that the `WordPressIntents` app extension which the variable
referred to is part of the `MANUALLY_MAINTAINED_STRINGS_FILES` `Hash`
which we pass to current iteration of the extractor action.