Use absolute paths when committing Fastlane metadata
The previous version used `Dir.glob('**/*.txt', base:
metadata_directory)`, but that didn't work for two reasons.

First, even if we use `base:`, the paths returned by `glob` are still
relative. From the docs:

> The optional base keyword argument specifies the base directory for
> interpreting relative pathnames instead of the current working
> directory. As the results are not prefixed with the base directory name
> in this case, you will need to prepend the base directory name if you
> want real paths.

See https://ruby-doc.org/core-2.7.4/Dir.html#method-c-glob .

Second, `Dir.glob` returns the files that match the given pattern, but
that excludes _deleted_ files. If a file is not there anymore, it can't
match the given pattern. This would have made the commit incomplete.

We could argue whether our code should delete the existing translations
when it doesn't find new ones. Even if we concluded that behavior is not
desirable, we'd still need to account for it locally until we update it
in the release toolkit. Otherwise, our local copies would remain dirty
after downloading the release notes metadata if no new ones are
available.