Widget: remove "Select Site Intent" from .strings
As discussed with @diegoreymendez in Slack, We're pretty sure that this string is never shown to the user, so instead of having polyglots scratching their head trying to  translate it (some already tried with translations that sounded more like "Select the intent for this site"… while we know that "Intent" here is the technical term Apple uses for .intentdefinition files and is not really a term that speaks to end users), better remove it altogether.

Also updated the hint for "Select Site" to try to provide a bit more context. We might decide to revisit those translator comments later to help better, but we want to land this ASAP to at least provide some comment instead of nothing before polyglots gets to those strings and translate them without the right context in mind.