Add all downloaded localized metadata before trying to commit them
I got this failure when running the `new_beta_release` lane:

```
[13:03:42]: ------------------------
[13:03:42]: --- Step: git_commit ---
[13:03:42]: ------------------------
[13:03:42]: $ git status /Users/gio/Developer/a8c/wpios/fastlane/metadata/\*\*/\*.txt --porcelain
[13:03:42]: ▸ ?? fastlane/metadata/tr/release_notes.txt
[13:03:42]: $ git commit -m Update\ WordPress\ metadata\ translations /Users/gio/Developer/a8c/wpios/fastlane/metadata/\*\*/\*.txt
[13:03:42]: ▸ On branch release/19.5
[13:03:42]: ▸ Your branch is ahead of 'origin/release/19.5' by 2 commits.
[13:03:42]: ▸ (use "git push" to publish your local commits)
[13:03:42]: ▸ Untracked files:
[13:03:42]: ▸ (use "git add <file>..." to include in what will be committed)
[13:03:42]: ▸ fastlane/metadata/tr/release_notes.txt
[13:03:42]: ▸ nothing added to commit but untracked files present (use "git add" to track)
+-----------------------------+----------------------+
|                    Lane Context                    |
+-----------------------------+----------------------+
| DEFAULT_PLATFORM            | ios                  |
| PLATFORM_NAME               | ios                  |
| LANE_NAME                   | ios new_beta_release |
| GIT_REPO_WAS_CLEAN_ON_START | true                 |
+-----------------------------+----------------------+
[13:03:42]: Exit status of command 'git commit -m Update\ WordPress\ metadata\ translations /Users/gio/Developer/a8c/wpios/fastlane/metadata/\*\*/\*.txt' was 1 instead of 0.
On branch release/19.5
Your branch is ahead of 'origin/release/19.5' by 2 commits.
  (use "git push" to publish your local commits)

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	fastlane/metadata/tr/release_notes.txt

nothing added to commit but untracked files present (use "git add" to track)
```

The `git_commit` docs for `path` say: The file(s) or directory(ies) you
want to commit. You can pass an array of multiple file-paths or
fileglobs "*.txt" to commit all matching files. The files already staged
but not specified and untracked files won't be committed.

I verified that `git commit` will fail if asked to commit when there's
nothing added to commit, which makes sense. At the same time, I'm
surprised by the failure because we pass `allow_nothing_to_commit`
`true` to `git_commit`, so this shouldn't happen. Maybe this is a hedge
case?

Regardless, we do want all those files to be committed.

The reason `fastlane/metadata/tr/release_notes.txt` was untracked is
because the `download_metadata.swift` script deletes the
`release_notes.txt` file for a given locale-key combination if it can't
find it, and we change the key (`glotPressWhatsNewKey`) every time we
make a new release.