Re-introduce `tVvJ9c` (intent description) key in `Sites.strings`
This was removed in 909570f8251b1265d2a5cccb6cc271284e28d8af and for
good reason: it's most likely not shown to the users.

But, App Store Connect gave us warnings upon submission that the
localized version of that string is missing.

> ITMS-90626: Invalid Siri Support - Localized description for custom intent: 'SelectSite' not found for locale: ar
> ITMS-90626: Invalid Siri Support - Localized description for custom intent: 'SelectSite' not found for locale: bg
> ITMS-90626: Invalid Siri Support - Localized description for custom intent: 'SelectSite' not found for locale: cs
>
> ... and so on for all locales other than en, en-AU, en-CA, and en-GB,
> even though those files don't have the key either (en-AU, en-CA, and
> en-GB are in fact empty)

This commit reintroduces the string with a message for the translators
to ignore it.