Remove direct link to the `.ipa` of Installable Build
The IPA that was downloadable from the PR comment was one built by our Enterprise Account and without restriction on the devices allowed, which means it could be downloaded by anyone without them requiring any specific permission to access it.

While the link to download the `.ipa` expired after 24h, this was still not great and could be considered a security issue, hence the complete removal of the link, forcing people who wants to install those Installable Builds to now always go thru AppCenter (which provides us a way to control permissions per user).