Spell out "one" instead of using numeric placeholder in singular string
This essentially reverts the change made in
8881d237ede89288f27d1f219bf177b36e1403fc.

While the version of the code from that commit, namely to have a numeric
placeholder two localized strings in `ReaderDetailLikesView` for a
number that would always result a "1" at runtime was technically correct
and in line with the String File Format references (see links below), in
practice some translators replaced it with a spelled out version.

Links:

- https://github.com/wordpress-mobile/WordPress-iOS/pull/16965#discussion_r682979305
- https://developer.apple.com/library/archive/documentation/MacOSX/Conceptual/BPInternational/StringsdictFileFormat/StringsdictFileFormat.html

For example, "%1$d Like" got translated to "O apreciere" in Romanian.
See
https://translate.wordpress.org/projects/apps/ios/dev/ro/default/?filters%5Bstatus%5D=either&filters%5Boriginal_id%5D=11994998&sort%5Bby%5D=translation_date_added&sort%5Bhow%5D=asc

As suggested by @AliSoftware,
https://github.com/wordpress-mobile/WordPress-iOS/pull/16965#discussion_r683345920
, we could "spell out" the numeral 1 to bypass confusion during
translation.

As a matter of fact, that's what the Romanian translator did.
Unfortunately, that's not okay in the codebase because the runtime
requires the placeholder to be there. I looked for crashes due to this
in Sentry but couldn't find any. That's either due to insufficient
Sentry searching skills on my end, or the possibility that there are
relatively few users with the app in the `ro` locale that the code path
never run.