Manually update `en.lproj/Localizable.strings` to remove long key
This is a followup to the localization fix implemented in
e6751413e4f2780336118850b33e7fa9098bb794.

Normally, we update the `en.lproj/Localizable.strings` file during code
freeze. In this instance, we decided to manually updated it to work
around a localization issue that our freshly re-enabled localization
linter highlighted. (The linter was re-enable in
cca74cbfe5ff36a6c00979110c49153fb814e963).

The alternative would have been a multiple release spanning process
where we first fixed the issues then we enabled the linter. It would
have dragged along for too long.

The plan from here is to:

> use the "Translation Memory" feature of GlotPress for that key,
> [the newly introduced `invite_people_invite_link_footer`] telling it
> to reuse the copy from the old key as the one for the new one.

More details in
https://github.com/wordpress-mobile/WordPress-iOS/pull/18348#issuecomment-1105492195