Run linter on app targets instead of appex ones
This changes the strategy to only have a single Script Build phase on the parent app target, rather than having one on each app extension target.

The initial rationale for calling the linter on each app extension target independently (and limiting the linter just to that target) was that it would ensure the linter would be run even if we only try to build the app extension target independently (e.g. selecting the scheme associated with that app extension when building). SO I thought it made more sense to make each target call the linter for their own.

But in practice, an app extension is never build / run without building the app target as well. Especially, if you look at each scheme corresponding to an app extension (e.g. the `WordPressStatsWidgets` scheme) you can see in the "Build" tab of the scheme that it will also build the `WordPress` app target anyway.

Given this, it ends up being (1) easier to maintain (e.g. if we add new appextension targets in the future), and (2) more performant / faster to just run the linter swift script once, from the app target, and have it scan the files for each of the appextension targets it finds in the `xcodeproj` in one go — rather than running it from scratch as many times as there are appextension targets in the project.

The performance boost we'll get from this new approach is because each invocation of the linter parses the whole `project.pbxproj` file (to find the list of targets then the list of files built by the targets). So only doing that once when we build the `WordPress` or `Jetpack` targets is more efficient than re-parsing the project file multiple times for each app extension target being built.