Stop resolving URLs of `PBXBuildFile` of groups
If a PBXBuildFile ends up pointing to a `PBXVariantGroup` or `XCVersionGroup` (as opposed to a `PBXFileReference`), this means to Xcode that all the `children` in that group will be built as part of the build phase.
For example, for a `PBXBuildFile` pointing to a `PBXVariantGroup` for `Localizable.strings`, Xcode will actually build all the `*.lproj/Localizable.strings` files that are `children` to that `PBXVariantGroup`.

In practice though, since we only care about linting `.string` and `.m` files and not `.strings` or `xcdatamodeld` and similar file-groupings, it's easier just not to bother with those and not even try to `resolveURL` on those since we'll skip them anyway when comes the call to `lint(…)`