Use current path for custom Jetpack metadata strings
Same rationale as e8469c7db4c5effc4749ddd8a3b4224aadbe367c:

These strings are the copy for the App Store screenshots and fallback
for release notes with no translation–currently unused.

Historically, these have lived in
`fastlane/appstoreres/jetpack_metadata/source/` and we forgot to account
for that when splitting up the `Fastfile` and updating the localization
tooling.

I don't know the consequences of adding custom `.txt` sources to the
folders `deliver` looks into. Given this work is happening in
`release/19.8` branch context, I'd rather do as little changes as
necessary, and save bigger restructuring for a PR against `trunk` on
which we can discuss and iterate without the release deadline looming
upon us.