Enforce decoding for objects of known `isa`
Before that change, if we found an object with a known `isa` value, we'd silently eat any error while parsing that known object.

Which could then lead to cryptic or misleading errors later down the line — e.g. an error about "failing to find the parent of an object" because we earlier silently failed to parsed that parent `PBXObject` as a typed object, instead of reporting the error at the point where the parsing of the `PBXObject` failed.

With that change if we find a known type but there's an error while trying to decode that known type, we'll now report it correctly.