Remove dispatch_once calls from mainContext and writerContext
The context objects are explicitly created during initialisation instead.

The dispatch_once calls have a side effect, which is only one intance of
each context object is created across _all_ instances of `ContextManager`
and its subclasses. To make it possible for all instances of
`ContextManager` to get valid value of `mainContext` and
`writerContext`, we'll need to remove the dispatch_once calls in their
property getters.