`git pull` before finalizing an hotfix or stable release
I got myself in a bad place today when finalizing WordPress iOS 19.9
because I didn't `git pull` before finalizing the build and I run the
automation from a machine other than the one where I launched the latest
beta build from.

The result was that the `git push` of the changes made by the release
finalization automation failed, because the remote had changes not
present locally.

I think what I experienced today is an edge case, and I'm usually
compulsively `git pull`-ing anyways, so there's low chance of this
happening again. Still, doesn't hurt to have the check in the codebase,
does it?

I decided to make the change here as opposed to the toolkit for two
reasons:

1. It's faster to ship it here, and I can always backport to the toolkit
   if we consider it a valuable addition (which is possible, as it would
   DRY and save us from forgetting to add it if we'll ever create new
   similar automations)
2. Recently, we've take to not perform Git operations via the toolkit,
   but let the calling `Fastfile` do that. This is in line with that
   pattern, even though the Git operations we've moved to `Fastfile` so
   far were all committing and pushing and, as far as I know, this is
   the first pulling.