Create missing sims only after Xcode invocation
Because when the CI VM starts, Xcode has never been launched yet, and the list of simulators returned by `xcrun simctl` is thus basically empty. But a default set of simulators is created by Xcode after the first launch of `xcodebuild` and invocation of `scan`.

So we need to create the missing simulators only **after** that first invocation of `xcodebuild`/`scan`, not before, to avoid creating simulators that would be also created by Xcode as part of the default set and end up with multiple simulators with the same device name, which would result in `capture_ios_screenshots` to fail to launch the UI test with a "The requested device could not be found because multiple devices matched the request." error otherwise.