Set WPStatsHomeWidgetsUserDefaultsSiteIdKey in StatsWidgetStore
Since WPStatsHomeWidgetsUserDefaultsSiteIdKey key is only used in the Stats Widget context it makes more sense to tie it to related Stats Widget events and not to bury it in more generic AccountService class.

We also have expectation for site data to be loaded after logInNotification and site data to be deleted after defaultSiteChanged notification, so setting user defaults key and then using it afterwards brings more transparency on what Stats widget depends on.