Fix app icon warning
Closes #18958

The app icon for the WordPress target was added to the membership of the Jetpack app. I don't think the Jetpack app needs access to the WordPress app icon, so I've removed its membership in this PR.

Looking at the code via git blame, it looks like this the WordPress app icon has been part of Jetpack since early 2021, so this isn't a new issue.