Revert "Update metadata strings for Stats Revamp"
This reverts commit bfaff682dab77c1dbad1c0768dc5c47d054dcf97.

I am reverting this because a change in this file would trigger a new
upload of the release notes to GlotPress, which might invalidate the
existing translation and create confusion with the localization team.

I want to discuss with them what's the best course of action before
moving forward with GlotPress changes. It might be simpler to remove the
translated line on the codebase hand, for example.