Manually restore WordPress `it` ASC metadata deleted due to HTTP error
The automation hit an HTTP error, likely the known HTTP 429 GlotPress
error (e.g. see
https://github.com/wordpress-mobile/release-toolkit/pull/402), and those
files were deleted.

The tooling is right in deleting the files when it can't find a
translation for them. But in this instance the translations were there,
we just had an issue in downloading them.

Notice that does not apply to the `release_notes.txt` file. There is no
translation available at this point in time.