Rename `branch_to_build` parameter into `branch`
This makes the lane parameter using the same name as the
`buildkite_trigger_build` action, avoiding confusion in the rare cases
we need to call it manually.

If you called the `trigger_release_build` or `trigger_beta_build` lane
without a parameter, you'll get prompted to pass a `branch` one, by the
underlying `buildkite_trigger_build` action.

But if you next time called the lane with a `branch` parameter, you'd
still get prompted to pass one, because the lane expected
`branch_to_build` , but this was not made clear by the messages Fastlane
prompted.

Rather than adding extra code for this, I thought aligning the
parameters name would sort out the issue. This leans a bit into the
convention over configuration realm, though. There's nothing stopping
future developers to reintroduce the same confusion by renaming the
parameter either at the lane or action level. Given the limited userbase
of this code and the relative slow pace at which it changes, I think
that's an acceptable tradeoff to have convenience for the time being.