Fix Jetpack app review URL
Ratings submitted for the Jetpack app were being submitted to the WordPress app's store listing. This happened because the app was using the incorrect URL to open the App Store app review screen.
While the app correctly defined the URL for the Jetpack app, it was then being overwritten at runtime by a URL fetched from an endpoint. This endpoint was only returning a URL for the WordPress app, so it overwrote the Jetpack app's URL with the WordPress app's one.

This PR mitigates the problem on the Jetpack app by skipping the endpoint request for the Jetpack app only. This doesn't change behavior on the WordPress app.
The impact of this change is that we can no longer remotely disable in-app requests for reviews. However, I don't recall this every being used.
The priority here is to restore the Rate Us functionality for the Jetpack app, so follow-up work can look for a more permanenent fix.