Use Fastlane's `AccountManager` to manage user in `register_new_device`
This removes a tiny bit of friction when addressing internal requests to
add a device to the Developer Portal, namely that if one forgets to call
the lane with their user in the environment, they'll get asked to insert
the credentials twice.

We can avoid that by pre-emptively configuring the users' credential for
Fastlane to use.

The logic and commentary are the same as WooCommerce iOS, see
https://github.com/woocommerce/woocommerce-ios/pull/7522 .

This begs the question of why not extracting the whole logic into a
dedicated new action. I guess that's something we might consider next
time we'll find ourselves editing this code.