Update lane running Fastlane `match` to run for both apps
We noticed in
https://github.com/wordpress-mobile/WordPress-iOS/pull/19328 that the
Jetpack provisioning profiles were not in sync with the information
stored in the Developer Portal.

While it's possible that whoever updated the Developer Portal (most
likely me) simply forgot to run the provisioning profiles automation in
this repo, even if they had it's likely the would have run into this
issue anyways because the public lane only updated the WordPress
profiles but the change was in the Jetpack ones.