Refresh project file after CocoaPods warning
Running `bundle exec pod update ...` in the previous commit,
b2b297c979465c14814ddaed5c5b636e7f08160d, printed a warning:

```
[!] `<PBXResourcesBuildPhase UUID=`FABB1FAA2602FC2C00C8785C`>` attempted to initialize an object with an unknown UUID. `FEC3B81A26C2915A00A395C7` for attribute: `files`. This can be the result of a merge and the unknown UUID is being discarded.
```

CocoaPods suggests there might have been a merge conflict not resolved
properly in the project file, which is something bound to happen from
time to time in a project with as many contributors as ours—and because
of the inconvenient way Xcode generates and manages the file.

Whenever I run into that kind of issue on my end, I add a new file then
remove it. That procedure is enough to make Xcode "refresh" the project
file, which usually gets rid of any dead reference, like the one
CocoaPods highlighted. And that's exactly what this commit tracks.

I also verified this changed by running the tests.