Remove redundant comments from `Podfile`
We don't need to encode the fact that when working with a
yet-to-be-released change in a library, which should happen only in a PR
context, we should point to the library's commit or branch where the
change took place.

Besides, comments are easy to ignore. We might get to implement so
static analysis to actually enforce this.