Handle migration from an older to a newer Jetpack app to support Stats widgets
After updating from an old version of Jetpack app that doesn't support Stats widgets to a newer version, required UserDefault flags and other information is not already saved and cannot be accessed by Jetpack Stats Widgets. Detect if these flags are not set and initialize widgets after the app is launched.