Use production version of internal pods for release 21.1
For the record, there is a new version of WordPressAuthenticator
available, 4.0.0, but it's a breaking change.

As far as I could see, all the work that went into it was driven by
WooCommerce and there is no need to update here immediately. As such,
I've decided to leave it and run the upgrade on a dedicated PR against
`trunk`.