Add step to load pods in cache before building and testing
It occurred to me, because of our steps parallelism, we can run in
situations where a commit that changes the `Podfile.lock` results in
multiple steps in the same CI build installing the pods from scratch.
For example, see this build where all three IPA deployment steps spend
15+ minutes in the CocoaPods part of the build because there was no
cached version of the pods available,
https://buildkite.com/automattic/wordpress-ios/builds/10636.

By introducing a step at the start of the pipeline, on which the
following steps depend upon, to install the pods we can make sure
parallel steps won't waste time building the same new version of the
Pods.