Reduce minElapstedTimeToRefresh to 1
We want to avoid calling API when getTimeline is called multiple times in a short period of time. For example when we reload cache and force reload widgets from the app.

However, anything more than that is excessive. If system decides that widget refresh is appropriate, we should allow that.