Use `.overCurrentContext` to present both Share and Save to Draft
Using `.overFullScreen` or leaving `.automatic` resulted in Share being
dismissable via swipe but that didn't restore Safari to the share menu,
leaving it instead blocked as if the presented VC was still there.

`.overCurrentContext` has no effect on the Save to Draft behavior,
leaving it the same as when using `.overFullScreen`.

On iPad, this setting result in both extensions not being full screen.
Otherwise, Share would have been full screen while Save to Draft
wouldn't.