[Stats Widget] Reduce the refresh interval for Stats widgets (#19694)
* Reduce refreshInterval from 60 to 30

For a widget the user frequently views, a daily budget typically includes from 40 to 70 refreshes. 30 minute refreshes fit into the budget while also provide more frequent refreshes for frequent users.

* Reduce minElapstedTimeToRefresh to 1

We want to avoid calling API when getTimeline is called multiple times in a short period of time. For example when we reload cache and force reload widgets from the app.

However, anything more than that is excessive. If system decides that widget refresh is appropriate, we should allow that.