Log unexpected HTTP requests instead of make the test case fail
I noticed there were failures caused by this line in CI jobs, but I
can't reproduce them on my mac.

It appears, on CI, requests sent by `ReaderTopicServiceRemote.fetchFollowedSites`
were captured by the stubs in `PeopleServiceTests`. I believe these
requests were triggered as a side effect of other unit tests. It just so
happened that they were caught by the stub which marked them as failure.

It's hard to avoid this kind of side effect, since the unit tests are running
in the host application and there is always going to be some code that's not
entirely related to the unit tests gets executed.