Address a "result of `flatMap` is unused" warning
From reading the code, it seems to me that the `flatMap` was used as a
way so conditionally run an operation based on whether the
`TimeInterval?` was not `nil`.

That seems redundant when we already have a `guard` that can do that
check for us and force a `return` if the input is `nil`.