Add simple test to load People screen
This relates to the issue reported in
https://github.com/wordpress-mobile/WordPress-iOS/issues/20112.

The test _doesn't actually catch the crash_.
When the People screen loads from the UI tests, it gets stuck in some
kind of loading loop where the animation continually restarts.

I decided to not investigate this further because the issue with the
loading loop is likely due to outdated or incomplete API request JSON
stubs and I don't have the time at this point to chase that down

I decide to keep the test anyway because it's a valuable starting point
to ensure that the people screen is available and interactive.