Redirect login-related support to forum
With this change, the app opens the support screen instead when the user initiates a support interaction. Previously, the app might first ask for their contact email. Now this is done through the support screen. This allows us to configure the support screen for the current app (WP or JP) and have all support interactions from the user flow through it.